Move all the dummy data from MapViewModel to a separate object file
